+++ src/repository/menu.repository.ts
@@ -1,0 +1,6 @@
+import AppDataSource from '../config/dataSource';
+import Menu from '../entity/menu.entity';
+
+const MenuRepository = AppDataSource.getRepository(Menu).extend({});
+
+export default MenuRepository;
