+++ src/repository/poll.repository.ts
@@ -1,0 +1,6 @@
+import AppDataSource from '../config/dataSource';
+import Poll from '../entity/poll.entity';
+
+const PollRepository = AppDataSource.getRepository(Poll).extend({});
+
+export default PollRepository;
