+++ src/repository/vote.repository.ts
@@ -1,0 +1,6 @@
+import AppDataSource from '../config/dataSource';
+import Vote from '../entity/vote.entity';
+
+const VoteRepository = AppDataSource.getRepository(Vote).extend({});
+
+export default VoteRepository;
