+++ src/service/poll.service.ts
@@ -1,6 +1,6 @@
 import Poll from '../entity/poll.entity';
 import CreatePollInput from '../type/poll/create.input';
-import { PollRepository } from '../repository/db.repository';
+import PollRepository from '../repository/poll.repository';
 import { InternalServerError } from '../util/customErrors';
 
 export default class PollService {
